Added some suggested changes
Removed the custom provider and custom window (html, ts and styles) as they are not needed to demonstrate security realms and add additional complexity to the sample (we can have other samples showing custom provider and window).

Updated the assets folder name to assets to match workspace starter.

Changed the iab command from send (which sends to a specific app) to publish (so it publishes to all runtimes that are listening on a topic).

Added kill command to package.json (just because it can be useful for a dev).

For launching manifest I try and get the root dynamically so that we can also publish this sample to github.io like we do for workspace-starter.

Outstanding: For some reason the taskbar icon isn't showing and I haven't had a chance to look into why.