Adam/how to/use security realms
An example by @adam-saland on using security realms across native, platform and window based OpenFin applications. 

This example includes a README, animated gifs to show the sample in action and two security realms to show that the messages are restricted to their respective realm  even though they share the same topic name and are based on the same codebase.